+++ base/src/test/kotlin/proguard/MarkerTest.kt
@@ -7,7 +7,10 @@
 
 package proguard
 
+import io.kotest.assertions.throwables.shouldNotThrowAny
 import io.kotest.core.spec.style.FreeSpec
+import io.kotest.matchers.shouldBe
+import io.kotest.matchers.shouldNotBe
 import proguard.io.ExtraDataEntryNameMap
 import proguard.mark.Marker
 import proguard.resources.file.ResourceFilePool
@@ -16,6 +19,7 @@ import proguard.testutils.KotlinSource
 import proguard.util.ProcessingFlags.DONT_OBFUSCATE
 import proguard.util.ProcessingFlags.DONT_OPTIMIZE
 import proguard.util.ProcessingFlags.DONT_SHRINK
+import proguard.util.kotlin.asserter.KotlinMetadataAsserter
 import testutils.asConfiguration
 import testutils.shouldHaveFlag
 import testutils.shouldNotHaveFlag
@@ -83,6 +87,94 @@ class MarkerTest : FreeSpec({
                     this shouldNotHaveFlag DONT_OPTIMIZE
                     this shouldNotHaveFlag DONT_SHRINK
                 }
+            }
+        }
+    }
+
+    "Given a Kotlin interface with default method implementation in the interface itself" - {
+        val (programClassPool, libraryClassPool) = ClassPoolBuilder.fromSource(
+            KotlinSource(
+                "Test.kt",
+                """
+                interface Test {
+                    fun foo() {
+                        TODO()
+                    }
+                }
+                """.trimIndent()
+            ),
+            kotlincArguments = listOf("-Xjvm-default=all")
+        )
+
+        // Run the asserter to ensure any metadata that isn't initialized correctly is thrown away
+        KotlinMetadataAsserter(Configuration()).execute(AppView(programClassPool, libraryClassPool))
+
+        "Then when marking" - {
+            val config = """
+            -keep class Test {
+                <methods>;
+            }
+            -keepkotlinmetadata
+            """.asConfiguration()
+            val marker = Marker(config)
+
+            "There should be no DefaultImpls class" {
+                programClassPool.getClass("Test\$DefaultImpls") shouldBe null
+            }
+
+            "An exception should not be thrown" {
+                shouldNotThrowAny {
+                    marker.execute(AppView(programClassPool, libraryClassPool))
+                }
+            }
+        }
+    }
+
+    "Given a Kotlin interface with default method implementation in compatibility mode" - {
+        val (programClassPool, libraryClassPool) = ClassPoolBuilder.fromSource(
+            KotlinSource(
+                "Test.kt",
+                """
+                interface Test {
+                    fun foo() {
+                        TODO()
+                    }
+                }
+                """.trimIndent()
+            ),
+            kotlincArguments = listOf("-Xjvm-default=all-compatibility")
+        )
+
+        // Run the asserter to ensure any metadata that isn't initialized correctly is thrown away
+        KotlinMetadataAsserter(Configuration()).execute(AppView(programClassPool, libraryClassPool))
+
+        "Then when marking" - {
+            val config = """
+            -keep class Test {
+                <methods>;
+            }
+            -keepkotlinmetadata
+            """.asConfiguration()
+            val marker = Marker(config)
+
+            "An exception should not be thrown" {
+                shouldNotThrowAny {
+                    marker.execute(AppView(programClassPool, libraryClassPool))
+                }
+            }
+
+            "There should be no DefaultImpls class" {
+                programClassPool.getClass("Test\$DefaultImpls") shouldNotBe null
+            }
+
+            val defaultImpls = programClassPool.getClass("Test\$DefaultImpls")
+
+            "The DefaultImpls class should be marked DONT_OPTIMIZE" {
+                defaultImpls.processingFlags shouldHaveFlag DONT_OPTIMIZE
+            }
+
+            "The default method should be marked DONT_OPTIMIZE" {
+                defaultImpls.findMethod("foo", null).processingFlags shouldHaveFlag DONT_OPTIMIZE
             }
         }
     }